requirements.txt
Flask>=3.0.0
yfinance>=0.2.31
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.3.1
textblob>=0.17.1
feedparser>=6.0.10
requests>=2.31.0
werkzeug>=3.0.0
beautifulsoup4>=4.12.2
dnspython>=2.4.2
